requirements.txt
numpy
wave
pyroomacoustics
pydub
scipy
matplotlib
ipykernel
PyYAML
librosa
